delete only when needed

+++ load_osm_file.py
@@ -23,8 +23,12 @@ def record(cursor, entry, identifier_of_region, timestamp):
         if len(data) == 1:
             present_already_timestamp = data[0][0]
             print("currently stored data has timestamp", present_already_timestamp, "our is", timestamp, (timestamp-present_already_timestamp), "seconds later")
-            # TODO do not always delete, only when needed
-            cursor.execute("DELETE FROM osm_data WHERE type = :type and id = :id", {'type': entry["osm_type"], 'id': entry["osm_id"]})
+            if timestamp > present_already_timestamp:
+                print("deleting older")
+                cursor.execute("DELETE FROM osm_data WHERE type = :type and id = :id", {'type': entry["osm_type"], 'id': entry["osm_id"]})
+            else:
+                print("so skipping")
+                return
 
         cursor.execute("INSERT INTO osm_data VALUES (:type, :id, :lat, :lon, :tags, :area_identifier, :download_timestamp, :validator_complaint)", {'type': entry["osm_type"], 'id': entry["osm_id"], 'lat': entry["lat"], 'lon': entry["lon"], "tags": json.dumps(entry["osm_tags"]), "area_identifier": identifier_of_region, "download_timestamp": timestamp, "validator_complaint": None})
